Changed user retreaval component
