+++ src/Router.tsx
@@ -7,11 +7,25 @@ import Newsletter from './components/Newsletter';
 import SignUp from './components/SignUp';
 import { createContext, useState } from 'react';
 import Profile from './components/Profile';
+import { useEffect } from 'react';
+import { getUserStatus } from './functions/DBGetUserStatus';
 
 export const UserContext = createContext<any>(null);
 export const LoadingContext = createContext<any>(false);
 const Router = () => {
-	const [user, setUser] = useState<null | Boolean>(null);
+	const [user, setUser] = useState<any>(null);
+	useEffect(() => {
+		const fetchUserStatus = async () => {
+			try {
+				const userStatus = await getUserStatus();
+				setUser(userStatus);
+			} catch (error) {
+				console.error('Error fetching user status:', error);
+				setUser(null);
+			}
+		};
+		fetchUserStatus();
+	}, []);
 	return (
 		<>
 			<UserContext.Provider value={{ user, setUser }}>
